Revert to 29f81c83e

+++ src/ResultWriter/PostProcessor.h
@@ -1,6 +1,8 @@
 #ifndef POST_PROCESSOR_H
 #define POST_PROCESSOR_H
 
+#include <vector>
+
 namespace seissol
 {
 
@@ -13,7 +15,7 @@ private:
     int m_numberOfVariables;
 public:
     PostProcessor (): m_numberOfVariables(0) {
-        for (unsigned int i = 0; i < 9; i++) {
+        for (size_t i = 0; i < 9; i++) {
             m_integrationMask[i] = false;
         }
     }
